Improve toggleGroup()
Now toggleGroup() works similarly to toggleItem(), by modifying the app's state data instead of the DOM, then allowing the JSX update the DOM based on the new data in state. Longer term, this is so that open/closed groups will also be saved to localStorage along with other user data for a more seamless user experience.