Update ListView to use Objects
The impetus of changing my data structure to use objects instead of arrays, ListView now also uses custom string IDs (instead of mutable numeric indices) to identify the exact location of a particular item in the JSON data, for the purpose of removing it from this filtered view. Eventually, this same system will also lend itself to changing the quantity of items in the List view.