Add Quantity to ListView
When an item is added from ItemsView, it gets a new property, called Quantity, which represents a total desired count for that item. In ListView, all numbers (yield and ingredient quantities) are multiplied by this number to tell the user how many ingredients they must collect in order to craft that many finished items. The "+" and "-" buttons  increment/decrement the quantity value, and the JSX renders the products of the quantity and corresponding original value. Additionally, when quantity is 1 and the user decrements, the item is removed from ListView. The quantity is retained in ItemsView if the user changes it and then removes the item from ListView.