Show Count of Pinned Items in Nav Button
I wanted to emulate a clever bit of UI in the game, where nav buttons display a "badge" with a count of "unread" items in that category. Mine will show the count of "pinned" items. My hope is this makes it very clear to the user that clicking items in the List view adds them to the Pinned view, and to then go check out that area.