Bigger Better PDA Graphic
Changes to image assets and CSS, representing a large visual update to the app's layout on medium and large devices, which seek to utilize more of the available screen space while more closely resembling the UI in the game.

As an added benefit, I used TinyPng.com to optimize the PNG file, which is roughly a third of the file size, while being physically larger than the original graphic.