Add Video Background
I sourced a good video of waves as seen from underwater by [raymond on Vecteezy](https://www.vecteezy.com/video/14398073-underwater-light-filters-down-through-blue-water-loop), and created an OpenShot project to edit it down in physical dimensions and length, and to convert it to a WEBM video. All of those project files are included, though I omitted the source video in lieu of a link to the download.