Export Item Images
These are straight out of Photoshop, using the built-in script, "Export Layers to Files." I modified that script so that it would use my layer names only, and not add numeric prefixes. I had to run it in batches, anything more than 50 or so layers at a time would yield solid black PNGs.

The next step is to optimize these using TinyPNG.