Refactor App to Root
One of the last stragglers in my conversion from Vue to React, the "root" DIV element where we mount our app has changed from ID `app` (the Vue default) to `root` (the React default). This fixes a visual issue where the app wasn't filling the height of the viewport on devices where the viewport was less than 1024px wide.